:white_check_mark: feat: 반복 아이콘 기능 추가 및 아이콘 유틸리티 테스트 처리
